# document-engine/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/v1/documents/{id}/history": {
      "get": {
        "tags": [
          "document-review-controller"
        ],
        "operationId": "history",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HistoryEntry": {
        "type": "object",
        "properties": {
          "subjectType": {
            "type": "string"
          },
          "subjectId": {
            "type": "string",
            "format": "uuid"
          },
          "fieldName": {
            "type": "string"
          },
          "groupKey": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "previousValue": {
            "type": "object"
          },
          "newValue": {
            "type": "object"
          },
          "reason": {
            "type": "string"
          },
          "decidedBy": {
            "type": "string",
            "format": "uuid"
          },
          "decidedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "HistoryResult": {
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string",
            "format": "uuid"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HistoryEntry"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}